Squashed commit of the following:
commit cd28b10c9b79a6d1d22cb7a3ac2b4b7335f4ac5f
Author: xRumax <167364@student.uwm.edu.pl>
Date:   Mon May 13 16:29:25 2024 +0200

    change of plans - not making a token/ reconstruction structure(services, routes, crud)

+++ app/crud/user.py
@@ -1,10 +1,14 @@
 from sqlalchemy.orm import Session
 from app.schemas.user import UserCreate, UserUpdate
 from app.models import user as models
-
+from app.utils.auth import hash_password, pwd_context
+from fastapi import HTTPException, status
 
 def create_user(db: Session, user: UserCreate):
-    db_user = models.User(**user.dict())
+    hashed_password = hash_password(user.password)
+    user_dict = user.dict()
+    user_dict["password"] = hashed_password
+    db_user = models.User(**user_dict)
     db.add(db_user)
     db.commit()
     db.refresh(db_user)
@@ -17,15 +21,25 @@ def get_user(db: Session, user_id: int):
     return db.query(models.User).filter(models.User.id == user_id).first()
 
 def update_user(db: Session, user: UserUpdate, user_id: int):
+    db_user = get_user(db=db, user_id=user_id)
+    if db_user is None:
+        raise HTTPException(status_code=404, detail="User not found")
     db.query(models.User).filter(models.User.id == user_id).update(user.dict())
     db.commit()
     return get_user(db=db, user_id=user_id)
 
 def delete_user(db: Session, user_id: int):
     db_user = get_user(db=db, user_id=user_id)
+    if db_user is None:
+        raise HTTPException(status_code=404, detail="User not found")
     db.delete(db_user)
     db.commit()
     return db_user
 
 def authenticate_user(db: Session, email: str, password: str):
-    return db.query(models.User).filter(models.User.email == email, models.User.password == password).first()
+    user = db.query(models.User).filter(models.User.email == email).first()
+    if not user:
+        return False
+    if not pwd_context.verify(password, user.password):
+        return False
+    return user
